+++ app/models/Category.scala
@@ -4,6 +4,6 @@ import play.api.libs.json._
 
 case class Category(id: Int, name: String)
 
-object Category{
-  implicit  val categoryFromat = Json.format[Category]
+object Category {
+  implicit val categoryFormat = Json.format[Category]
 }
