+++ conf/application.conf
@@ -100,7 +100,6 @@ play.filters {
   # Enabled filters are run automatically against Play.
   # CSRFFilter, AllowedHostFilters, and SecurityHeadersFilters are enabled by default.
   enabled += filters.ExampleFilter
-
   # Disabled filters remove elements from the enabled list.
   #disabled += filters.ExampleFilter
 }
@@ -318,11 +317,12 @@ play.filters {
 # Evolutions allows database scripts to be automatically run on startup in dev mode
 # for database migrations. You must enable this by adding to build.sbt:
 #
-# libraryDependencies += evolutions
+libraryDependencies += evolutions
 #
 play.evolutions {
   # You can disable evolutions for a specific datasource if necessary
-  db.default.enabled = true
+  enabled = true
+  db.default.enabled =true
 }
 
 ## Database Connection Pool
@@ -373,6 +373,7 @@ db {
   # https://www.playframework.com/documentation/latest/Highlights25#Logging-SQL-statements
   #default.logSql=true
 }
-slick.db.default.driver =  "slick.driver.SQLiteDriver$"
-slick.db.default.db.driver =  "org.sqlite.JDBC"
-slick.db.default.url =  "jdbc:sqlite:shop:sqlite"
+
+slick.dbs.default.driver="slick.driver.SQLiteDriver$"
+slick.dbs.default.db.driver="org.sqlite.JDBC"
+slick.dbs.default.db.url="jdbc:sqlite:ebiznes.sqlite"
